+++ ldda_vol/graph.h
@@ -40,24 +40,24 @@ public:
 
     void init_lemon();
 
-    void update_single_weight(long,long);
-    void update_all_weights(vector<long>);
+    void update_single_weight(long,double);
+    void update_all_weights(vector<double>);
 
     void lemon_delete_edge(long);
     vector<long> lemon_contract_edge(long);
     bool lemon_graph_modified;
 
     bool mst();
-    long mst_weight;
+    double mst_weight;
     vector<bool> mst_vector;
     double mst_runtime;
 
-    pair<bool,long> mst_probing_var(long, bool);
+    pair<bool,double> mst_probing_var(long, bool);
     double probe_runtime;
 
     // volume only: maximum weight spanning tree primal bound for LDDA
     bool maxst();
-    long maxst_weight;
+    double maxst_weight;
     vector<bool> maxst_vector;
 
 private:
@@ -74,9 +74,9 @@ private:
     vector< list<long> > adj_list;
 
     // edge list
-    vector<long> s;        // terminal node 1
-    vector<long> t;        // terminal node 2
-    vector<long> w;        // edge weight
+    vector<long> s;           // terminal node 1
+    vector<long> t;           // terminal node 2
+    vector<double> w;    // edge weight
 
     // adjacency matrix storing edge indexes
     bool using_matrix;
@@ -88,19 +88,19 @@ private:
     vector<ListGraph::Node> lemon_vertices;
     vector<ListGraph::Edge> lemon_edges;
     ListGraph::EdgeMap<long> *lemon_edges_inverted_index;
-    ListGraph::EdgeMap<long> *lemon_weight;
+    ListGraph::EdgeMap<double> *lemon_weight;
 
     bool lemon_test_adj(ListGraph &, ListGraph::Node &, ListGraph::Node &);
     ListGraph::Edge lemon_test_adj_getting_edge(ListGraph &,
                                                 ListGraph::Node &,
                                                 ListGraph::Node &);
     vector<ListGraph::Edge> lemon_parallel_edges_if_contract(ListGraph &,
-                                                            ListGraph::EdgeMap<long> &,
+                                                            ListGraph::EdgeMap<double> &,
                                                             ListGraph::Node &,
                                                             ListGraph::Node &);
 
     // volume only: negative/opposite edge weights
-    ListGraph::EdgeMap<long> *opposite_weights;
+    ListGraph::EdgeMap<double> *opposite_weights;
 };
 
 #endif
